+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Fixed
 - Fixed issue #4 in FirstReactionMethod
 
+### Changed
+- stocal.examples.type_rules now does what it claims to do
+
 
 ## [1.1.1] - 2018-03-21